backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyMySQL==1.1.1
python-jose==3.3.0
requests==2.32.3
